+++ vonixcore-1.19.2-fabric-quilt-forge-template/common/src/main/java/network/vonix/vonixcore/VonixCore.java
@@ -32,7 +32,7 @@ public class VonixCore {
 
     public static final String MODID = "vonixcore";
     public static final String MOD_NAME = "VonixCore";
-    public static final String VERSION = "1.3.2";
+    public static final String VERSION = "1.4.0";
 
     public static final Logger LOGGER = LoggerFactory.getLogger(MOD_NAME);